+++ PRO-controlPRO/src/js/app-metadata.js
@@ -37,9 +37,9 @@
 
   var metadata = Object.freeze({
     appName: 'Semilla App Web',
-    version: '1.10.3',
-    build: '2026-08-05-167',
-    releasedAt: '2026-08-05T07:55:00-05:00',
+    version: '1.10.4',
+    build: '2026-08-05-168',
+    releasedAt: '2026-08-05T15:24:00-05:00',
     applicationId: applicationId,
     applicationPath: applicationId === 'root' ? '/' : '/' + applicationId + '/',
     applicationBaseUrl: applicationBaseUrl,